refactor: rotation done with quaternion and transform updated rather than repalced
